+++ R/generate_pattern.R
@@ -1,19 +1,17 @@
-#' Generate a CAT pattern
+#' Generate a CAT patterns
 #' 
-#' Generate a CAT pattern given various inputs. Returns a character or numeric vector 
-#' with length equal to the test size, depending on whether a \code{choices} input was supplied.
+#' Generate a CAT pattern given various inputs. Returns a character vector or numeric matrix 
+#' (depending on whether a \code{df} input was supplied) with columns equal to the test size and
+#' rows equal to the number of rows in \code{Theta}. For simulation studies, supplying a 
+#' \code{Theta} input with more than 1 row will run independent CAT session when passed to 
+#' \code{mirtCAT()}.
 #' 
 #' @param mirt_object single group object defined by the \code{mirt} package
 #'
 #' @param Theta a numeric vector indicating the latent theta values for a single person
 #' 
-#' @param choices a list of character vectors signifying the possible choices for each item.
-#'   If NULL, a numeric vector is returned indicating the selected category for each item (where
-#'   0 is the lowest possible category)
-#' 
-#' @param item_answers (optional) a character vector indicating which of the options in 
-#'   \code{choices} should be considered the 'correct' answer. This is required for item types that 
-#'   dichotomously score items (e.g., multiple choice items scored with the 3PL model)
+#' @param df (optional) data.frame object containing questions, options, and scoring
+#'   keys. See \code{\link{mirtCAT}} for details
 #' 
 #' @export generate_pattern
 #' @author Phil Chalmers \email{rphilip.chalmers@@gmail.com}
@@ -23,31 +21,41 @@
 #' \dontrun{
 #' 
 #' # return real response vector given choices and (optional) answers 
-#' pat <- generate_pattern(mod, Theta = 0, choices = choices, item_answers=answers)
-#' # mirtCAT(questions, mirt_object=mod, local_pattern = pat)
+#' pat <- generate_pattern(mod, Theta = 0, df=df)
+#' # mirtCAT(df, mirt_object=mod, local_pattern = pat)
 #' 
-#' # generate numeric pattern observed in dataset used to define mod
+#' # generate single pattern observed in dataset used to define mod
 #' pat2 <- generate_pattern(mod, Theta = 0)
 #' # mirtCAT(mirt_object=mod, local_pattern = pat2)
+#'
+#' # generate multiple patterns to be analyzed independently 
+#' pat3 <- generate_pattern(mod, Theta = matrix(c(0, 2, -2), 3))
+#' # mirtCAT(mirt_object=mod, local_pattern = pat3)
 #' 
 #' }
-generate_pattern <- function(mirt_object, Theta, choices = NULL, item_answers = NULL){
+generate_pattern <- function(mirt_object, Theta, df = NULL){
+    fn <- function(p, ns) sample(1L:ns, 1L, prob = p) - 1
     nitems <- ncol(mirt_object@Data$data)
-    pattern <- matrix(0L, 1L, nitems)
     if(!is.matrix(Theta)) Theta <- matrix(Theta, 1L)
-    if(is.null(choices)){
+    N <- nrow(Theta)
+    pattern <- matrix(0L, N, nitems)
+    if(is.null(df)){
         K <- mirt_object@Data$K
         for(i in 1L:nitems){
             ii <- extract.item(mirt_object, i)
             P <- probtrace(ii, Theta)
-            uniq <- 0L:(K[i]-1)
-            pattern[i] <- sample(uniq, 1L, prob = P)
+            pattern[,i] <- apply(P, 1L, fn, ns = ncol(P))
         }
-        return(as.numeric(pattern + mirt_object@Data$mins))
+        return(t(t(pattern) + mirt_object@Data$mins))
+    } else {
+        choices <- df[,grepl('Option', colnames(df))]
+        item_answers <- df[,grepl('Answer', colnames(df))]
+        if(is.matrix(item_answers))
+            stop('Only one correct answer is supported when drawing data')
     }
     ret <- character(nitems)
     K_1 <- do.call(c, lapply(choices, length)) - 1L
-    has_item_answers <- !is.null(item_answers)
+    has_item_answers <- length(item_answers) > 0L
     if(has_item_answers)
         K_1[!is.na(item_answers)] <- 1L
     for(i in 1L:nitems){
@@ -57,7 +65,7 @@ generate_pattern <- function(mirt_object, Theta, choices = NULL, item_answers = 
         pattern[i] <- sample(uniq, 1L, prob = P)
         if(has_item_answers){
             ret[i] <- if(pattern[i] == 1L) item_answers[i] else
-                sample(choices[[i]][!choices[[i]] %in% item_answers[i]], 1L)
+                sample(as.character(choices[i, ][!choices[i, ] %in% item_answers[i]]), 1L)
         } else {
             ret[i] <- choices[[i]][pattern[i]+1L]
         }